requirements.txt
Flask
flask_sqlalchemy
python-dotenv
sqlalchemy
mysql